[fix]

See merge request s13-final/S13P31A103!94

Former-commit-id: 873a4378dd305966881337e81440bfa5911a32a8

+++ backend/spring-server/src/main/java/com/heungbuja/song/dto/SongGameData.java
@@ -34,7 +34,7 @@ public class SongGameData implements Serializable {
 
     // 가공된 데이터
     private SectionInfo sectionInfo;
-    private Integer bpm;
+    private Double bpm;
     private Double duration;
 
     // 동작 타임라인
